Changes regarding ‘dbh’ units
Changes in ‘tree.detection’ function for obtaining ‘dbh’ values in cm (instead of m) and, as a consequence, changes in ‘distance.sampling’, ‘estimation.plot.size’, ‘metrics.variables’, and ‘simulations’ functions to adapt them to this unit conversion.
In addition, element ‘tree.list.tls’ included in ‘Rioja.data’ dataset was also modified to contain ‘dbh’ values in cm (instead of m).